@@ -13,15 +13,16 @@ apply plugin: 'com.android.application'
 apply from: rootProject.file('gradle/codequality.gradle')
 
 android {
-    compileSdkVersion 19
+    compileSdkVersion 30
     buildToolsVersion "22.0.1"
 
     defaultConfig {
         versionCode 101
         versionName "1.9.1"
 
-        minSdkVersion 10
-        targetSdkVersion 19
+        minSdkVersion 14
+        targetSdkVersion 30
+
 
         buildConfigField 'String', 'BUILD_TAG', '"' + getBuildTag() + '"'
         buildConfigField 'String', 'OWM_API_KEY', '"' + getOpenWeatherMapApiKey() + '"'
@@ -30,7 +31,7 @@ android {
         def buildSuffix = getBuildSuffix(versionName, versionCode)
         applicationVariants.all { variant ->
             def file = variant.outputs[0].outputFile
-            variant.outputs[0].outputFile = new File(file.parent, file.name.replace(".apk", "-" + buildSuffix + ".apk"))
+       // variant.outputs[0].outputFile = new File(file.parent, file.name.replace(".apk", "-" + buildSuffix + ".apk"))
         }
     }
 
@@ -48,6 +49,7 @@ android {
     buildTypes {
         release {
             minifyEnabled true
+
             proguardFile 'proguard-project.txt'
             if (project.hasProperty('signingKeyStoreFile')) {
                 signingConfig signingConfigs.release
@@ -106,10 +108,18 @@ repositories {
 }
 
 dependencies{
-    compile 'com.actionbarsherlock:actionbarsherlock:4.4.0@aar'
-    compile 'com.android.support:support-v4:19.1.0'
-    compile 'com.google.code.gson:gson:2.3.1'
-    compile group: 'net.time4j', name: 'time4j-android', version: '4.2-2018g'
+    //implementation 'com.actionbarsherlock:actionbarsherlock:4.4.0@aar'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    //testCompile 'junit:junit:4.12'
+    //implementation 'com.android.support:support-v4:28.0.0'
+    //implementation 'com.android.support:cardview-v7:28.0.0'
+    //implementation 'com.android.support:appcompat-v7:28.0.0'
+    //implementation 'com.android.support:design:28.0.0'
+    //implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.code.gson:gson:2.8.7'
+    implementation group: 'net.time4j', name: 'time4j-android', version: '4.2-2018g'
 }
 
 
