@@ -23,6 +23,8 @@ import android.util.AttributeSet;
 import android.widget.Checkable;
 import android.widget.FrameLayout;
 
+import ch.corten.aha.worldclock.R;
+
 /**
  * A frame to create checkable list items in Android Versions older than Honeycomb.
  */
@@ -52,10 +54,10 @@ public class CheckableFrameLayout extends FrameLayout implements Checkable {
     public void setChecked(boolean checked) {
         mChecked = checked;
         if (checked) {
-            Drawable drawable = getContext().getResources().getDrawable(com.actionbarsherlock.R.drawable.abs__list_activated_holo);
+            Drawable drawable = getContext().getResources().getDrawable(R.drawable.abs__list_activated_holo);
             setBackgroundDrawable(drawable);
         } else {
-            setBackgroundDrawable(new ColorDrawable(android.R.color.transparent));
+            setBackgroundDrawable(new ColorDrawable(getResources().getColor(R.color.transparent)));
         }
     }
 
